Create generic `StripeBottomSheetLayout` and use in Payments
