+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/PaymentOptionsActivity.kt
@@ -10,13 +10,13 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.lifecycle.ViewModelProvider
-import com.stripe.android.common.ui.BottomSheet
-import com.stripe.android.common.ui.rememberBottomSheetState
+import com.stripe.android.common.ui.ElementsBottomSheetLayout
 import com.stripe.android.paymentsheet.ui.BaseSheetActivity
 import com.stripe.android.paymentsheet.ui.PaymentSheetFlowType.Custom
 import com.stripe.android.paymentsheet.ui.PaymentSheetScreen
 import com.stripe.android.paymentsheet.utils.applicationIsTaskOwner
 import com.stripe.android.uicore.StripeTheme
+import com.stripe.android.uicore.elements.bottomsheet.rememberStripeBottomSheetState
 import kotlinx.coroutines.flow.filterNotNull
 
 /**
@@ -53,7 +53,7 @@ internal class PaymentOptionsActivity : BaseSheetActivity<PaymentOptionResult>()
             StripeTheme {
                 val isProcessing by viewModel.processing.collectAsState()
 
-                val bottomSheetState = rememberBottomSheetState(
+                val bottomSheetState = rememberStripeBottomSheetState(
                     confirmValueChange = { !isProcessing },
                 )
 
@@ -65,7 +65,7 @@ internal class PaymentOptionsActivity : BaseSheetActivity<PaymentOptionResult>()
                     }
                 }
 
-                BottomSheet(
+                ElementsBottomSheetLayout(
                     state = bottomSheetState,
                     onDismissed = viewModel::onUserCancel,
                 ) {
