+++ stripe-ui-core/res/anim/stripe_transition_fade_in.xml
@@ -2,5 +2,5 @@
 <alpha xmlns:android="http://schemas.android.com/apk/res/android"
     android:duration="@android:integer/config_shortAnimTime"
     android:interpolator="@android:anim/decelerate_interpolator"
-    android:fromAlpha="1"
-    android:toAlpha="0" />
+    android:fromAlpha="0"
+    android:toAlpha="1" />
